# [EWT-2995] Optimize the performance of RTree search